Added config and commands for Talos Linux.
This works by patching the default talos configuration with only specific changes that need to be made to the overall config. Currently, most of this is network specific.

Additionally, we have configuration for _specific_ node types (currently only my RPi 4 with 4GB) which will make that configuration unique to the specific machine. The other patches are more general and can/should be applied to all configs.