First setup docs/recipes must handle insecure/bootstrap.
When first setting up a node from a freshly flashed image, we need to use `--insecure` to apply the initial conifg. This is now possible with a flag that can be set on the command.

Also, a freshly flashed node must bootstrap etcd, because otherwise it'll wait forever. Added a simply command to do this, too.