Setup Litestream init/sidecar containers for SQLite replication.
My setup first creates an "emptyDir" mount to start with a clean stage. The initContainer then restores the DB from it's latest version (if available). Then, both the Application and the sidecar container start. While the application writes to the DB file, the sidecar container contiously replicates to the S3 URL target.

This is following their official Kubernetes guide. The module abstracts away most of it, so it's easily reusable. Also, Litestream supports multiple DBs to be replicated and multiple targets to replicate to, so if need be this could be expanded to be more powerful.