Allow setting user/group permissions for SQLite DB
Application images may use any user _inside_ the image to read/write to/from the SQLite DB. These are subject to "normal" Linux directory permissions. This now allows customizing these.

The way it works is that the volume containing the Database is mounted with the set GID. Then, we make sure to start _both_ Litestream containers with the same GID and UID (if set). This ensures that if the database is restored, it's created with the correct UID/GID automatically. The Lightstream containers don't need to run as root.