EXPERIMENT: Get Gotify Application Token via API Request.
We're using the generic `http` Terraform Provider to make a POST request to the Gotify REST API to create a new application and then use the `token` for setting the Diun config input.

This works, and with the `livecycle` thing it actually waits until the deployment is ready and a request can be made to the API. It has two problems in it's current for:

1. The provider is not smart and the request isn't idempotent: It's repeated on every apply. This means we set a new token on each apply and they pile up in Gotify.
2. This won't work when setting up from scratch, because we use the external DNS name to reach the service. We'd need to make the request to the IP directly and set the `Host` HTTP header manually.

Will look into other methods of archiving this.