Ensure Gotify requests will work without DNS setup.
This will allow a deployment from scratch to work without needing to pause to change the machines DNS settings. We send the request to the static IP address and manually set the `Host` as expected by the Ingress. We also need to wait for both the Ingress (for request routing) and Gotify (for the actual API) to be up.

Will verify this setup with a disaster recovery soon.