Monitoring: Added fluent-bit to look for Litestream problems.
We're doing this by ingesting all our Kubernetes Container logs, then filtering only the `litestream` containers. Currently, we're routing based on the log-level:

- info/warn/error are printed to stdout in the fluent-bit container
- warn/error are additionally available to be re-routed elsewhere

Currently everything is simply logged to stdout, but for warn/error we want to send messages to gotify in the future. Figuring out how to do that now...