Made the SQLite replication optional for Web Apps.
Also moved everything to do with replication into a single object. This makes working with a group of values easier, especially when _some_ of them are optional and others are required. Terraform has good support for this.

The only required change was a `moved` instruction to make the optional ConfigMap work without re-creating it.