Web service always exposed on port 80.
This gets rid of a problem where once the service is applied _once_, the `targetPort` isn't specifically set and will always be what it was _initially_. This can cause issues when the apps `exposed_port` is changed later on.

Additionally, this makes the cluster-internal usage of services easier, since no port is needed when all web services listen on port 80.