Use Gotify with slack-incoming-webhook plugin.
This is my own plugin, so we're switching to the generated multi-arch container with gotify and the plugin bundled together.

Gotifies default SQLite database folder is `/app/data/db.sqlite`. The default plugin directory is `/app/data/plugins`. Mounting the `emptyDir` volume for the database to `/app/data` then overlays our plugin in the bundled container. To avoid this, I have moved the database into it's own deeper sub-path, so that the plugin isn't overlayed and found as expected.