Send Litestream issues to Gotify.
To get this working, we're using FluentBit to analyze the Kubernetes logs and look for warning/errors from any litestream pods. If these are found, the specific issue is sent via the Slack output to Gotify. The "slack-incoming-webhook" plugin for Gotify then accepts the Slack message and adds it to a Gotify application channel.

TLS note: The `slack` output in FluentBit has a hard-coded check that fails startup if the Slack Webhook URL does not have the HTTPS schema in it. Since we're not sending any traffic to slack/over the public internet, we're side-stepping this validation: We use the HTTPS schema but also explicitly target port 80. Then, we disable any TLS validation.