FIX: Ensure no two instances of Litestream apps.
Litestream does NOT support being written to by more than one instance. Also, an app that is shutting down might flush it's database while stopping. With the default "rolling" strategy, a new app instance could already be running and the init-container might not restore the database from the newest snapshot.

Changing the strategy to "replace" makes sure that the old instance is down entirely before starting a new one. This ensures that no two instances write to Litestream and that the init-container restores from the newest Litestream snapshot.