Diun: Check all tags.
This is necessary because we use specific image tags for apps. It will look for all new tags in a specific repository, order them by SemVer and pick the newest one. If a newer one than the one currently in use is available, a notification is sent to Gotify.

In theory at least. I haven't seen any of these notifications yet..