squash! draft deployment of watch-your-lan.
It's unclear if this will apply as-is. It's also unclear if it works without hard-coding device specific stuff. Perhaps we need a way to inject this into Terraform (vars?). Additionally, it's unclear if this should be a "normal" web-app as the others, since it does more in the background.

Lots of questions at the moment. Postponing adding this for now...