squash! DRAFT deployment of Actual budgeting software.
This does not save its data in an SQLite database. There are a few interesting projects now that have this limitation and don't really work with my current setup. It would be interesting to explore Talos Linux local storage options (we can think about replication when more than one node is in the cluster) and perhaps just have regular backup jobs along with it (per deployment?).