Dashboard: Don't allow selecting search results via mouse.
This usually gets in the way if the mouse is _somewhere_ on the screen and I simply start typing. The mouse will often hover a result somewhere down the list and that makes it auto-selected. I then need to move the mouse to not open the hovered selection via the enter key.

Removing all mouse-pointer events from the `dialog` container disables hovering interactions. It also shows the pointer with an icon to indicate interaction is disabled. This is good enough for my use-case for now.