Made Bookmark groups more flexible in Dashboard.
This allows adding arbitrary groups of bookmarks to the dashboard. The groups are no longer pre-determined.

The drawback of the current solution is that due to Terraform `for` convention, all keys in the map are sorted alphabetically. This makes it hard to influence the order in which the bookmark groups show up in the dashboard. I have tried to find some more clever naming to fix this for now.