Added `CronJob` to run `replica-verify` command.
This will automatically be setup if a web-app uses SQLite replication. It creates both a `CronJob` and a corresponding Healthchecks.io check for each one. The status and its logs are reported there.

The times at which the checks run are defined as arguments to the module. I've configured it now to give each one 20min to complete (far too much time) so they stay out of each other's way.