Added info command and links to renew Certificates.
There are _two_ certificates involved: The one in the "kubeconfig" file for `kubectl` access and the one in the "talosconfig" file for `talosctl` access.
If "kubeconfig" expires, run `just kubeconfig` to renew it. This only works if the "talosconfig" one is still valid.
If the "talosconfig" is expired, too, regenerate it from the `secrets.yml` file (exists locally and in 1password).

Taken from https://github.com/siderolabs/talos/discussions/9457